start-storage

+++ internal/models/models.go
@@ -5,5 +5,5 @@ type Url struct {
 	Url       string `db:"url"`
 	ShortUrl  string `db:"short_url"`
 	CreatedAt int    `db:"created_at"`
-	UpdatedAt int    `db:"updated_at"`
+	UpdatedAt int64  `db:"updated_at"`
 }
